refactor: move utils functions to separate files
