+++ src/pages/info-page/components/line-chart/line-chart.tsx
@@ -5,15 +5,19 @@ import { VictoryArea, VictoryAxis, VictoryChart } from "victory";
 import { useAppDispatch, useAppSelector } from "../../../../hooks";
 import { fetchHistoryData } from "../../../../redux/actions/get-coins-asynk-thunk";
 
-import { HistoryCoinType } from "../../../../types/coin.__types__";
+import {
+  createRechartsObj,
+  tickFormatX,
+  tickFormatY,
+} from "../../utils/line-chart";
+
+type LineRechartsProps = {
+  coinId: string;
+};
 
 type DataRechartType = {
   name: string;
   price: string;
-};
-
-type LineRechartsProps = {
-  coinId: string;
 };
 
 export const LineRecharts: React.FC<LineRechartsProps> = ({ coinId }) => {
@@ -21,27 +25,13 @@ export const LineRecharts: React.FC<LineRechartsProps> = ({ coinId }) => {
   const { historyCoin } = useAppSelector((state) => state.coins);
   const dispatch = useAppDispatch();
 
-  function createRechartsObj(rechartsObj: HistoryCoinType[]) {
-    const arrTenDays = rechartsObj.slice(-30);
-    const arr = arrTenDays.map((elem) => {
-      const objDate = new Date(elem.date);
-      return {
-        name: `${objDate.getDate()}.${
-          objDate.getMonth() + 1
-        }.${objDate.getFullYear()}`,
-        price: Number(elem.priceUsd).toFixed(2),
-      };
-    });
-    setDataRecharts(arr);
-  }
-
   useEffect(() => {
     dispatch(fetchHistoryData({ id: coinId, interval: "d1" }));
   }, [dispatch, coinId]);
 
   useEffect(() => {
     if (coinId) {
-      createRechartsObj(historyCoin);
+      setDataRecharts(createRechartsObj(historyCoin));
     }
   }, [historyCoin, coinId]);
 
@@ -61,25 +51,9 @@ export const LineRecharts: React.FC<LineRechartsProps> = ({ coinId }) => {
             interpolation='natural'
           />
 
-          <VictoryAxis
-            tickFormat={(t, index) => {
-              const tickStep = 5;
-              if (index % tickStep === 0) {
-                return t;
-              }
-              return "";
-            }}
-          />
+          <VictoryAxis tickFormat={tickFormatX} />
 
-          <VictoryAxis
-            dependentAxis
-            tickFormat={(tick, index, ticks) => {
-              if (index === 0 || tick !== ticks[index - 1]) {
-                return tick;
-              }
-              return "";
-            }}
-          />
+          <VictoryAxis dependentAxis tickFormat={tickFormatY} />
         </VictoryChart>
       )}
     </div>
