+++ src/pages/main-page/components/table/types/types.ts
@@ -1,0 +1,10 @@
+export type OneCoinTableType = {
+    changePercent24Hr: string;
+    id: string;
+    marketCapUsd: string;
+    name: string;
+    priceUsd: string;
+    rank: string;
+    symbol: string;
+    vwap24Hr: string;
+  };
